+++ packages/vue/src/Hen.vue
@@ -1,4 +1,4 @@
-<!-- GENERADO por scripts/build.ts — no editar a mano. ·  Fuente: packages/core/src + design-tokens.json · AviVet Icons v0.3.1 -->
+<!-- GENERADO por scripts/build.ts — no editar a mano. ·  Fuente: packages/core/src + design-tokens.json · AviVet Icons v0.4.0 -->
 <script setup lang="ts">
 withDefaults(defineProps<{ size?: number | string; title?: string }>(), { size: 24 });
 </script>
@@ -22,6 +22,6 @@ withDefaults(defineProps<{ size?: number | string; title?: string }>(), { size: 
     focusable="false"
   >
     <title v-if="title">{{ title }}</title>
-    <path d="M26 20C24 12 14 12 12 19C12 23 16 24 21 24C19 31 24 48 34 50C45 50 53 42 52 31C40 30 32 27 26 20Z"/><path d="M23.5 14.5C24 10.5 20 10 20 13.5C20 9 16 9 16 13C16 10.5 14 11 14 14.5"/><path d="M12.5 16.5L6 19L12.5 21.5Z"/><circle cx="17" cy="17.5" r="1.25" fill="currentColor" stroke="none"/><path d="M46 40C42.85 33.7 34.45 31.6 27.1 35.8"/><path d="M50 30C52.5 28 55.5 25.5 57.5 21.5"/><path d="M50 31.5C52.5 30 55.5 28.5 58 25.5"/><path d="M50 33C52.5 32 55.5 31.5 58 29.5"/><path d="M38.7 49.5L38.7 56.5"/><path d="M29.6 48.7L29.6 56.5"/>
+    <path d="M24 16C19 11 13 14 11 22C11 25 13 27 16 28C19 29 20 31 19 34C16 42 22 50 31 51C40 52 46 49 47 44C48 39 50 34 49 31C48 28 41 25 35 26C31 25 28 20 24 16Z"/><path d="M24 16.5C24 11 20 10.5 20 15C20 9.5 15.5 9 15.5 14C15.5 10 12 11 12.5 16.5"/><path d="M12 25.5C9.5 27.5 10 31 12.5 30.5Z"/><path d="M11.5 19.5L5 22L11.5 24.5Z"/><circle cx="17.5" cy="18" r="1.25" fill="currentColor" stroke="none"/><path d="M43 42C39 34 31 30 24 32C25 39 34 43 43 42Z"/><path d="M39.5 42.5C35 37 30 34.5 25 36"/><path d="M47 28C52 24 56 17 57 10C54 15 50 21 46 25Z"/><path d="M48.5 32C54 30 57.5 24 58.5 17C55.5 21 51.5 26 47 29Z"/><path d="M49 36C54.5 35 58.5 31 60 26C56.5 28 52 31 48 33.5Z"/><path d="M35.5 51L35.5 55.5"/><path d="M32 59L35.5 55.5L39 59"/><path d="M35.5 55.5L35.5 59.5"/><path d="M26 49.5L26 55.5"/><path d="M22.5 59L26 55.5L29.5 59"/><path d="M26 55.5L26 59.5"/>
   </svg>
 </template>